MUTEX guard for UART TX
Multiple Priority tasks racing for TX DR buffer
When low priority owns the mutex, and high pr task recognises and wait
for it. If a medium priority task wants to start and preempt the low
priority, low priority keeps work to the completion of the job and high
priority acquires the mutex and then medium priority job can get the
CPU or resource.

smart priority inheritence in source code.
Need to study this intellegent method.