FEAT: Event Group
Traceable

Event Group are another inter-task synchronization objects that offer
additional features compared to semaphores:

    With Event Group, you can have a task waiting for a combination of
events to occur (not only one), each coming from a separate task
    Several tasks can wait for the same combination of events and get
unblocked altogether when that combination occur. Of course, the first
waiting task to get active is the one of highest priority level.

Using an event group can replace several binary semaphores, thus saving
memory.

Basically, a Event Group is a 16 or 32 bit register where each bit is a
flag that represents the state On/Off of a user-defined event. By using
usual masking techniques, a task can wait for a number of flags to be
set with either an AND-type or an OR-type filter. In addition, the
waiting task may choose or not to reset the awaited event flags after
reading.